reorganised all character insertion components into a dedicated folder
