+++ scribemaster/src/components/Character/AssetCreation/Insertion/EntityCard.tsx
